Update handling of FMA option for CLang
DBA requires the FMA operation on Intel to work properly.

However, the flag to enable it does not exist on Apple silicon.

Unfortunately, the FMA Intel binary will not run in Rosetta, but
you should really be using a native binary instead.

Ideally, a universal binary could work -- with FMA on Intel, but
not on Apple silicon.